[SPTFC-7] created a MainHeader component
