+++ package.json
@@ -22,6 +22,7 @@
     "@clerk/clerk-react": "^4.26.1",
     "@phosphor-icons/react": "^2.0.10",
     "@radix-ui/react-context-menu": "^2.1.5",
+    "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-scroll-area": "^1.0.4",
     "@radix-ui/react-slider": "^1.1.2",
     "@tanstack/react-query": "^5.8.4",
